+++ src/processorhandler.h
@@ -327,7 +327,7 @@ private:
 
   QFutureWatcher<void> m_runWatcher;
   bool m_stopRunningFlag = false;
-  bool m_clockFinished = true;
+  std::mutex m_clockLock;
 
   /**
    * @brief To avoid excessive UI updates due to things relying on